requirements.txt
langchain==0.0.274
chromadb==0.4.11
urllib3==2.0.4
PyMuPDF==1.23.1
python-dotenv==1.0.0
unstructured==0.10.8
markdown==3.5.1
python-pptx==0.6.23
extract-msg==0.45.0
tabulate==0.9.0
pandoc==2.3
pypandoc==1.11
tqdm==4.66.1
sentence_transformers==2.2.2
discord.py==2.3.2
aiohttp==3.9.3
dataclasses-json==0.5.14
